Renamed package structure
went from `peckb1.xxx` to `com.github.peckb1.xxx`
