+++ examples/src/main/java/com/github/peckb1/examples/auto/Gorg.java
@@ -1,0 +1,14 @@
+package com.github.peckb1.examples.auto;
+
+import com.github.peckb1.processor.AutoJackson;
+
+import java.util.Optional;
+
+@AutoJackson
+public interface Gorg {
+
+    Integer getAge();
+    String getName();
+    Optional<Gorg> getChild();
+
+}
