+++ examples/src/main/java/com/github/peckb1/examples/auto/Muppet.java
@@ -1,0 +1,6 @@
+package com.github.peckb1.examples.auto;
+
+
+public enum Muppet {
+    KERMIT, MISS_PIGGY, FOZZIE, GONZO, SCOOTER, ROWLF
+}
