+++ examples/src/main/java/com/github/peckb1/examples/auto/fraggles/Mokey.java
@@ -1,0 +1,12 @@
+package com.github.peckb1.examples.auto.fraggles;
+
+import com.github.peckb1.examples.auto.Fraggle;
+import com.github.peckb1.processor.AutoJackson;
+import com.github.peckb1.processor.Named;
+
+@AutoJackson
+public interface Mokey extends Fraggle {
+
+    @Named("radishes") int getRadishesPicked();
+
+}
