--- examples/src/main/java/peckb1/examples/auto/Gorg.java
@@ -1,14 +1,0 @@
-package peckb1.examples.auto;
-
-import peckb1.processor.AutoJackson;
-
-import java.util.Optional;
-
-@AutoJackson
-public interface Gorg {
-
-    Integer getAge();
-    String getName();
-    Optional<Gorg> getChild();
-
-}
