--- examples/src/main/java/peckb1/examples/auto/Muppet.java
@@ -1,6 +1,0 @@
-package peckb1.examples.auto;
-
-
-public enum Muppet {
-    KERMIT, MISS_PIGGY, FOZZIE, GONZO, SCOOTER, ROWLF
-}
