+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>peckb1</groupId>
+    <groupId>com.github.peckb1</groupId>
     <artifactId>autojackson</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>pom</packaging>
