+++ processor/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>peckb1</groupId>
+        <groupId>com.github.peckb1</groupId>
         <artifactId>autojackson</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
