+++ processor/src/main/java/com/github/peckb1/processor/NoTypeEnum.java
@@ -1,0 +1,7 @@
+package com.github.peckb1.processor;
+
+/**
+ * A Simple enumeration used internally to signify that the user did not provide
+ * a custom type inside their {@link AutoJackson#type()} variable.
+ */
+enum NoTypeEnum { }
