--- processor/src/main/java/peckb1/processor/NoTypeEnum.java
@@ -1,7 +1,0 @@
-package peckb1.processor;
-
-/**
- * A Simple enumeration used internally to signify that the user did not provide
- * a custom type inside their {@link AutoJackson#type()} variable.
- */
-enum NoTypeEnum { }
